Add files via upload
Added the Python file containing the revised code with notes on my initial mistakes. The code correctly performs the simple check to support how probabilities sum to 1. Additionally, the code correctly computes the examples given in the ClassicalRWonCircle and QuantumRWonCircle .ipynb files. In my revised code, added as Muhammad_Revised_Plot.py, the parameters tested are for the case when L = 101, and when t ranges from 0 to 1. This commit also includes a screenshot of the output plot for this test.